+++ .storybook/main.ts
@@ -1,4 +1,5 @@
 import type { StorybookConfig } from '@storybook/nextjs';
+import path from 'path';
 
 const config: StorybookConfig = {
   stories: ['./docs/**/*.mdx', '../**/*.stories.@(js|jsx|ts|tsx)'],
@@ -30,9 +31,10 @@ const config: StorybookConfig = {
     /* Aliases */
     if (!config.resolve) config.resolve = {};
     if (!config.resolve.alias) config.resolve.alias = {};
-
-    (config.resolve.alias as { [index: string]: string })['next/navigation'] =
-      require.resolve('../__mocks__/next/navigation');
+    config.resolve.alias = {
+      ...config.resolve.alias,
+      'next/navigation': path.resolve('./__mocks__/next/navigation'),
+    };
 
     /* Plugins */
     if (!config.plugins) config.plugins = [];